requirements.txt
faiss_cpu
gradio
numpy
ollama
